refactor training
